docs(ecommerce): move ecommerce example to HMR
