+++ docgen/src/ecommerce.html
@@ -5,5 +5,4 @@ name: ecommerce
 nav_groups:
   - core
 ---
-
-<div id="e-commerce-example"></div>
+<div id="app"></div>
